Introduce new lexer-based sql normalization.
This improves performance quite a bit in the general case and, more importantly, fixes several edge cases in regex handling like stack overflows and very slow loopbacks.
Adds tests showing that these edge cases don't affect the lexer implementation.
Fixes a subtle muzzle bug involving primitive types and helper classes that was in the way of this work.

Adjusts decorators to not emit the db.statement tag if normalization fails (which should be quite rare).